+++ ImageFeed/ImagesList/ImagesListViewController.swift
@@ -1,0 +1,46 @@
+import UIKit
+
+final class ImagesListViewController: UIViewController {
+    @IBOutlet private var tableView: UITableView!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        tableView.delegate = self
+        tableView.dataSource = self
+//        tableView.register(
+//            ImagesListCell.self,
+//            forCellReuseIdentifier: ImagesListCell.reuseIdentifier
+//        )
+        
+        tableView.rowHeight = 200
+    }
+}
+
+extension ImagesListViewController: UITableViewDelegate {
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) { }
+}
+
+extension ImagesListViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 1
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: ImagesListCell.reuseIdentifier, for: indexPath)
+        
+        guard let imageListCell = cell as? ImagesListCell else {
+            // TODO Тут лучше логировать ошибку
+            return UITableViewCell()
+        }
+        
+        configCell(for: imageListCell)
+        
+        return imageListCell
+    }
+    
+    private func configCell(for cell: ImagesListCell) {
+        cell.layer.cornerRadius = 16
+        cell.layer.masksToBounds = true
+    }
+}
